refactor(all): move TS specific implementation to separate dir
