+++ src/BaseGenerator.ts
@@ -4,8 +4,11 @@ import {IApiController, IApiData, IApiMethod} from "./interfaces";
 import {GeneratorHelpers} from "./GeneratorHelpers";
 
 export abstract class BaseGenerator {
-  protected abstract generateInterfaces();
-  protected abstract generateControllers();
+  public abstract generate();
+  protected abstract generateModels();
+  protected abstract generateApis();
+  protected abstract bundle();
+
   protected apiData: IApiData;
   protected emptyApiControllers: IApiController[];
   protected apiControllers: IApiController[];
@@ -17,18 +20,5 @@ export abstract class BaseGenerator {
     this.emptyApiControllers = GeneratorHelpers.getEmptyApiControllers(this.apiData.paths);
     this.apiMethods = GeneratorHelpers.getApiMethods(this.apiData.paths);
     this.apiControllers = GeneratorHelpers.getApiControllers(this.emptyApiControllers, this.apiMethods);
-  }
-
-  public generate() {
-    // todo: move this to TypeScriptGenerator.ts this belongs there. Not here.
-    const interfaceString = this.generateInterfaces();
-    const controllersString = this.generateControllers();
-    const sdk = interfaceString + controllersString;
-    fs.writeFile("sdk/TypeScript/src/sdk.ts", sdk, (err) => {
-      if (err) throw err;
-      console.log("The file has been saved!");
-    });
-    fs.writeFileSync("sdk/TypeScript/package.json", fs.readFileSync(path.resolve("./src/package.stub")));
-    fs.writeFileSync("sdk/TypeScript/tsconfig.json", fs.readFileSync(path.resolve("./src/tsconfig.stub")));
   }
 }
