+++ src/TypeScriptGenerator/InterfaceGenerator.test.ts
