+++ src/TypeScriptGenerator/stubs/docblock.stub
