+++ src/TypeScriptGenerator/stubs/interface.stub
