+++ src/TypeScriptGenerator/stubs/package.stub
