+++ src/TypeScriptGenerator/stubs/service_method.stub
