+++ src/TypeScriptGenerator/stubs/shop_service.stub
