+++ src/TypeScriptGenerator/stubs/tsconfig.stub
